sync time
Sync the time of the web client with the one of the server.

Sync the time between the watch and the phone and the time of
the phone with the time on the server.

This is done to corelate the data from the web client with the
data from the watch