collect-sensors: update ui
Add Reset, Pause and Start button.

Collecting starts when Start is pressed.

Reset clears all data and leaves the collecting paused. It can't
be pressed while syncing time as pressing it triggers time syncing
and it would cause a loop.

Pause interrupts the collecting of the data until it's resumed
or sent.

When the memory is full the app pauses the collecting until Reset
or Send ar pressed.

After it's used reset the value of the id to an empty string.