add knn classifier with dtw metric
Add a k-nearest-neighbors classifier that directly takes as input
the signal and it uses the dynamic time warping metric.

The data is converted into lists that do not have the id for easier
access.

The code obtaining the classes for the pressed keys is moved to a
separate function to avoid duplicated code.